+++ src/pubmed/author.rs
@@ -58,8 +58,7 @@ impl AuthorName {
         }
     }
 
-    // not used, consider deleting?
-    /// Get the name as an `AU`.
+    #[allow(dead_code)]
     pub fn as_au(&self) -> Cow<'_, str> {
         if self.full {
             let initials = self.first_initials();